# backend/src/docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"MessageResponse": {
				"description": "Generic success envelope",
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {}
				},
				"required": [
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RegisterRequest": {
				"description": "Body for POST /auth/register",
				"properties": {
					"email": {
						"type": "string",
						"description": "User email address"
					},
					"password": {
						"type": "string",
						"description": "Minimum 8 characters"
					},
					"role": {
						"type": "string",
						"description": "One of: STUDENT | COMPANY | ACADEMIC_SUPERVISOR | SITE_SUPERVISOR | ADMIN"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password",
					"role",
					"firstName",
					"lastName"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LoginRequest": {
				"description": "Body for POST /auth/login",
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ForgotPasswordRequest": {
				"description": "Body for POST /auth/forgot-password",
				"properties": {
					"email": {
						"type": "string"
					}
				},
				"required": [
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResetPasswordRequest": {
				"description": "Body for POST /auth/reset-password",
				"properties": {
					"token": {
						"type": "string"
					},
					"password": {
						"type": "string",
						"description": "Minimum 8 characters"
					}
				},
				"required": [
					"token",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"EvaluationBody": {
				"properties": {
					"enrollmentId": {
						"type": "string"
					},
					"stage": {
						"type": "string",
						"description": "MIDTERM | FINAL"
					},
					"punctuality": {
						"type": "number",
						"format": "double",
						"description": "1–5"
					},
					"communication": {
						"type": "number",
						"format": "double",
						"description": "1–5"
					},
					"technicalSkills": {
						"type": "number",
						"format": "double",
						"description": "1–5"
					},
					"teamwork": {
						"type": "number",
						"format": "double",
						"description": "1–5"
					},
					"initiative": {
						"type": "number",
						"format": "double",
						"description": "1–5"
					},
					"professionalism": {
						"type": "number",
						"format": "double",
						"description": "1–5"
					},
					"strengths": {
						"type": "string"
					},
					"areasForImprovement": {
						"type": "string"
					},
					"generalComments": {
						"type": "string"
					},
					"recommendForHire": {
						"type": "boolean"
					}
				},
				"required": [
					"enrollmentId",
					"stage"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SelfAssessmentBody": {
				"properties": {
					"enrollmentId": {
						"type": "string"
					},
					"stage": {
						"type": "string",
						"description": "MIDTERM | FINAL"
					},
					"skillsDeveloped": {
						"type": "number",
						"format": "double",
						"description": "1–5"
					},
					"goalsMet": {
						"type": "number",
						"format": "double",
						"description": "1–5"
					},
					"supervisorSupport": {
						"type": "number",
						"format": "double",
						"description": "1–5"
					},
					"workEnvironment": {
						"type": "number",
						"format": "double",
						"description": "1–5"
					},
					"overallExperience": {
						"type": "number",
						"format": "double",
						"description": "1–5"
					},
					"achievements": {
						"type": "string"
					},
					"challenges": {
						"type": "string"
					},
					"lessonsLearned": {
						"type": "string"
					},
					"futureGoals": {
						"type": "string"
					}
				},
				"required": [
					"stage"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "Internship Monitoring & Connecting System API",
		"version": "1.0.0",
		"description": "Full-stack platform bridging students, companies, and supervisors throughout the internship lifecycle.",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "API Support"
		}
	},
	"paths": {
		"/api/auth/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"description": "Register a new user account.\r\nAn email verification link is sent after successful registration.",
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RegisterRequest"
							}
						}
					}
				}
			}
		},
		"/api/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"description": "Authenticate with email and password.\r\nReturns a JWT bearer token.",
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginRequest"
							}
						}
					}
				}
			}
		},
		"/api/auth/verify-email": {
			"get": {
				"operationId": "VerifyEmail",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"description": "Verify a user's email address using the token sent by email.",
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "token",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/auth/forgot-password": {
			"post": {
				"operationId": "ForgotPassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"description": "Request a password reset link.\r\nAlways returns 200 to prevent email enumeration.",
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ForgotPasswordRequest"
							}
						}
					}
				}
			}
		},
		"/api/auth/reset-password": {
			"post": {
				"operationId": "ResetPassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"description": "Reset password using the token from the reset email.",
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ResetPasswordRequest"
							}
						}
					}
				}
			}
		},
		"/api/auth/me": {
			"get": {
				"operationId": "GetMe",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"description": "Get the currently authenticated user's profile.",
				"summary": "Get current user",
				"tags": [
					"Auth"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			}
		},
...
		"/api/evaluations": {
			"post": {
				"operationId": "SubmitEvaluation",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"description": "Submit a supervisor evaluation for an enrolled student.\r\nRequires ACADEMIC_SUPERVISOR or SITE_SUPERVISOR role.",
				"tags": [
					"Evaluations"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/EvaluationBody"
							}
						}
					}
				}
			}
		},
		"/api/evaluations/enrollment/{enrollmentId}": {
			"get": {
				"operationId": "GetEvaluations",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"description": "Get all evaluations for a specific enrollment",
				"tags": [
					"Evaluations"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "enrollmentId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/evaluations/{id}/approve": {
			"patch": {
				"operationId": "ApproveEvaluation",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"description": "Approve a submitted evaluation (Admin only)",
				"tags": [
					"Evaluations"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/evaluations/self-assessment": {
			"post": {
				"operationId": "SubmitSelfAssessment",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"description": "Submit a student self-assessment (Student only)",
				"tags": [
					"Evaluations"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SelfAssessmentBody"
							}
						}
					}
				}
			}
		},
		"/api/evaluations/self-assessment/me": {
			"get": {
				"operationId": "GetMySelfAssessments",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"description": "Get all self-assessments for the authenticated student",
				"tags": [
					"Evaluations"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	],
...
}